// packages/twenty-apps/internal/resend/src/types/update-template.dto.ts
import type { EmailsField } from 'src/utils/to-emails-field';

export type UpdateTemplateDto = {
  name: string;
  alias: string;
  status: string;
  fromAddress: EmailsField;
  subject: string;
  replyTo: EmailsField;
  htmlBody: string;
  textBody: string;
  resendUpdatedAt: string;
  publishedAt: string | null;
};